# openapi/variable-definitions-internal.yml
openapi: 3.0.1
info:
  title: Internal Variable Definitions Administration API
  description: |2+

    ## Introduction

    Variable Definitions are centralized definitions of concrete variables which are typically present in multiple datasets. Variable Definitions support standardization of data and metadata and facilitate sharing and joining of data by clarifying when variables have an identical definition.

    ## Maintenance of Variable Definitions
    This API allows for creation, maintenance and access of Variable Definitions.

    ### Ownership
    Creation and maintenance of variables may only be performed by Statistics Norway employees representing a specific Dapla team, who are defined as the owners of a given Variable Definition. The team an owner represents must be specified when making a request through the `active_group` query parameter. All maintenance is to be performed by the owners, with no intervention from administrators.

    ### Status
    All Variable Definitions have an associated status. The possible values for status are `DRAFT`, `PUBLISHED_INTERNAL` and `PUBLISHED_EXTERNAL`.

    #### Draft
    When a Variable Definition is created it is assigned the status `DRAFT`. Under this status the Variable Definition is:

    - Only visible to Statistics Norway employees.
    - Mutable (it may be changed directly without need for versioning).
    - Not suitable to refer to from other systems.

...

    #### Published Internal
...

...

...

    #### Published External
...

...

...

...

...

...

...
servers:
...
security: []
...
paths:
  /vardok-migration:
    get:
      tags:
      - Data Migration
      summary: Get a list of all vardok and vardef id mappings
      description: Get a list of all vardok and vardef id mappings
      operationId: getVardokVardefMapping
      responses:
        "200":
          description: OK response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VardokVardefIdPairResponse"
              examples:
                Vardef vardok id list:
                  value: |-
                    [
                        {
                        "vardok_id": "1607",
                        "vardef_id": "wypvb3wd"
                    }
                    ]
      security:
      - labid_token: []
  /vardok-migration/{vardef-id}:
    get:
      tags:
      - Data Migration
      summary: Get a vardok id by vardef id.
      description: Get a vardok id by vardef id.
      operationId: getVardokByVardefId
      parameters:
      - name: Vardef ID
        in: path
        description: The Vardef ID.
        required: true
        schema:
          type: string
        examples:
          Vardef id:
            value: wypvb3wd
          Not found:
            value: Xy9_7-Az
      - name: vardef-id
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: https://opensource.zalando.com/restful-api-guidelines/models/problem-1.0.1.yaml#/Problem
              examples:
                Not found:
                  value:
                    type: about:blank
                    status: 404
                    detail: Not found
                    parameters: {}
        "200":
          description: OK response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VardokIdResponse"
              examples:
                Vardef id:
                  value:
                    vardok_id: "1607"
      security:
      - labid_token: []
  /vardok-migration/{vardok-id}:
    get:
      tags:
      - Data Migration
      summary: Get a variable definition by vardok id.
      description: Get a variable definition by vardok id.
      operationId: getVardefByVardokId
      parameters:
      - name: Vardok ID
        in: path
        description: The ID of the definition in Vardok.
        required: true
        schema:
          type: string
        examples:
          Vardok id:
            value: "1607"
          Not found:
            value: "9999"
      - name: vardok-id
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: https://opensource.zalando.com/restful-api-guidelines/models/problem-1.0.1.yaml#/Problem
              examples:
                Not found:
                  value:
                    type: about:blank
                    status: 404
                    detail: Not found
                    parameters: {}
        "200":
          description: OK response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompleteView"
              examples:
                Vardok id:
                  value:
                    id: wypvb3wd
                    patch_id: 1
                    name:
                      en: Country Background
                      nb: Landbakgrunn
                      nn: Landbakgrunn
                    short_name: landbak
                    definition:
                      en: "Country background is the person's own, the mother's or\
                        \ possibly the father's country of birth. Persons without\
                        \ an immigrant background always have Norway as country background.\
                        \ In cases where the parents have different countries of birth\
                        \ the mother's country of birth is chosen. If neither the\
                        \ person nor the parents are born abroad, country background\
                        \ is chosen from the first person born abroad in the order\
                        \ mother's mother, mother's father, father's mother, father's\
                        \ father."
                      nb: "For personer født i utlandet, er dette (med noen få unntak)\
                        \ eget fødeland. For personer født i Norge er det foreldrenes\
                        \ fødeland. I de tilfeller der foreldrene har ulikt fødeland,\
                        \ er det morens fødeland som blir valgt. Hvis ikke personen\
                        \ selv eller noen av foreldrene er utenlandsfødt, hentes landbakgrunn\
                        \ fra de første utenlandsfødte en treffer på i rekkefølgen\
                        \ mormor, morfar, farmor eller farfar."
                      nn: "For personar fødd i utlandet, er dette (med nokre få unntak)\
                        \ eige fødeland. For personar fødd i Noreg er det fødelandet\
                        \ til foreldra. I dei tilfella der foreldra har ulikt fødeland,\
                        \ er det fødelandet til mora som blir valt. Viss ikkje personen\
                        \ sjølv eller nokon av foreldra er utenlandsfødt, blir henta\
                        \ landsbakgrunn frå dei første utenlandsfødte ein treffar\
                        \ på i rekkjefølgja mormor, morfar, farmor eller farfar."
                    classification_reference: "91"
                    unit_types:
                    - "01"
                    - "02"
                    subject_fields:
                    - he04
                    contains_special_categories_of_personal_data: true
                    variable_status: DRAFT
                    measurement_type: "01"
                    valid_from: 2003-01-01
                    external_reference_uri: https://www.ssb.no/a/metadata/conceptvariable/vardok/1919/nb
                    comment:
                      nb: Fra og med 1.1.2003 ble definisjon endret til også å trekke
                        inn besteforeldrenes fødeland.
                      nn: Fra og med 1.1.2003 ble definisjon endret til også å trekke
                        inn besteforeldrenes fødeland.
                      en: "As of 1 January 2003, the definition was changed to also\
                        \ include the grandparents' country of birth."
                    related_variable_definition_uris:
                    - https://example.com/
                    owner:
                      team: team-a
                      groups:
                      - team-a-developers
                    contact:
                      title:
                        en: Division for population statistics
                        nb: Seksjon for befolkningsstatistikk
                        nn: Seksjon for befolkningsstatistikk
                      email: s320@ssb.no
                    created_at: 2024-06-11T08:15:19.038Z
                    created_by: ano@ssb.no
                    last_updated_at: 2024-06-11T08:15:19.038Z
                    last_updated_by: ano@ssb.no
      security:
      - labid_token: []
    post:
      tags:
      - Data Migration
      summary: Create a variable definition from a VarDok variable definition.
      description: Create a variable definition from a VarDok variable definition.
      operationId: createVariableDefinitionFromVarDok
      parameters:
      - name: vardok-id
        in: path
        description: The ID of the definition in Vardok.
        required: true
        schema:
          type: string
        examples:
          Migrate Vardok:
            value: "1607"
      responses:
        "201":
          description: Successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompleteView"
              examples:
                Migrate Vardok:
                  value:
                    id: wypvb3wd
                    patch_id: 1
                    name:
                      en: Country Background
                      nb: Landbakgrunn
                      nn: Landbakgrunn
                    short_name: landbak
                    definition:
                      en: "Country background is the person's own, the mother's or\
                        \ possibly the father's country of birth. Persons without\
                        \ an immigrant background always have Norway as country background.\
                        \ In cases where the parents have different countries of birth\
                        \ the mother's country of birth is chosen. If neither the\
                        \ person nor the parents are born abroad, country background\
                        \ is chosen from the first person born abroad in the order\
                        \ mother's mother, mother's father, father's mother, father's\
                        \ father."
                      nb: "For personer født i utlandet, er dette (med noen få unntak)\
                        \ eget fødeland. For personer født i Norge er det foreldrenes\
                        \ fødeland. I de tilfeller der foreldrene har ulikt fødeland,\
                        \ er det morens fødeland som blir valgt. Hvis ikke personen\
                        \ selv eller noen av foreldrene er utenlandsfødt, hentes landbakgrunn\
                        \ fra de første utenlandsfødte en treffer på i rekkefølgen\
                        \ mormor, morfar, farmor eller farfar."
                      nn: "For personar fødd i utlandet, er dette (med nokre få unntak)\
                        \ eige fødeland. For personar fødd i Noreg er det fødelandet\
                        \ til foreldra. I dei tilfella der foreldra har ulikt fødeland,\
                        \ er det fødelandet til mora som blir valt. Viss ikkje personen\
                        \ sjølv eller nokon av foreldra er utenlandsfødt, blir henta\
                        \ landsbakgrunn frå dei første utenlandsfødte ein treffar\
                        \ på i rekkjefølgja mormor, morfar, farmor eller farfar."
                    classification_reference: "91"
                    unit_types:
                    - "01"
                    - "02"
                    subject_fields:
                    - he04
                    contains_special_categories_of_personal_data: true
                    variable_status: DRAFT
                    measurement_type: "01"
                    valid_from: 2003-01-01
                    external_reference_uri: https://www.ssb.no/a/metadata/conceptvariable/vardok/1919/nb
                    comment:
                      nb: Fra og med 1.1.2003 ble definisjon endret til også å trekke
                        inn besteforeldrenes fødeland.
                      nn: Fra og med 1.1.2003 ble definisjon endret til også å trekke
                        inn besteforeldrenes fødeland.
                      en: "As of 1 January 2003, the definition was changed to also\
                        \ include the grandparents' country of birth."
                    related_variable_definition_uris:
                    - https://example.com/
                    owner:
                      team: team-a
                      groups:
                      - team-a-developers
                    contact:
                      title:
                        en: Division for population statistics
                        nb: Seksjon for befolkningsstatistikk
                        nn: Seksjon for befolkningsstatistikk
                      email: s320@ssb.no
                    created_at: 2024-06-11T08:15:19.038Z
                    created_by: ano@ssb.no
                    last_updated_at: 2024-06-11T08:15:19.038Z
                    last_updated_by: ano@ssb.no
        "400":
          description: Bad request.
          content:
            application/problem+json:
              schema:
                $ref: https://opensource.zalando.com/restful-api-guidelines/models/problem-1.0.1.yaml#/Problem
              examples:
                Bad request:
                  value:
                    type: about:blank
                    status: 400
                    detail: "Failed to convert argument [example] for value [null]\
                      \ due to: Error deserializing type: example"
                Constraint violation:
                  value:
                    suppressed: []
                    parameters: {}
                    type: https://zalando.github.io/problem/constraint-violation
                    title: Constraint Violation
                    status: 400
                    violations:
                    - field: classificationReference
                      message: Code incorrect is not a valid classification id
      security:
      - labid_token: []
...
components:
  schemas:
    CompleteView:
      required:
      - contact
      - contains_special_categories_of_personal_data
      - created_at
      - created_by
      - definition
      - id
      - last_updated_at
      - last_updated_by
      - name
      - owner
      - patch_id
      - short_name
      - subject_fields
      - unit_types
      - valid_from
      - variable_status
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the variable definition.
          format: "^[a-zA-Z0-9-_]{8}$"
        patch_id:
          type: integer
          description: Integer identifying a patch of a variable definition.
          format: int32
          example: 1
        name:
          $ref: "#/components/schemas/LanguageStringType"
        short_name:
          type: string
          description: Recommended short name. Must be unique within an organization.
        definition:
          $ref: "#/components/schemas/LanguageStringType"
        classification_reference:
          type: string
          description: ID of a classification or code list from Klass. The given classification
            defines all possible values for the defined variable.
          nullable: true
        unit_types:
          type: array
          description: "A list of one or more unit types, e.g. person, vehicle, household.\
            \ Must be defined as codes from https://www.ssb.no/klass/klassifikasjoner/702."
          items:
            type: string
        subject_fields:
          type: array
          description: A list of subject fields that the variable is used in. Must
            be defined as codes from https://www.ssb.no/klass/klassifikasjoner/618.
          items:
            type: string
        contains_special_categories_of_personal_data:
          type: boolean
          description: "True if variable instances contain particularly sensitive\
            \ information. Applies even if the information or identifiers are pseudonymized.\
            \ Information within the following categories are regarded as particularly\
            \ sensitive: Ethnicity, Political alignment, Religion, Philosophical beliefs,\
            \ Union membership, Genetics, Biometrics, Health, Sexual relations, Sexual\
            \ orientation"
        variable_status:
          $ref: "#/components/schemas/VariableStatus"
        measurement_type:
          type: string
          description: "Type of measurement for the variable, e.g. length, volume,\
            \ currency. Must be defined as codes from https://www.ssb.no/klass/klassifikasjoner/303"
          nullable: true
        valid_from:
          type: string
          description: The variable definition is valid from this date inclusive
          format: date
        valid_until:
          type: string
          description: The variable definition is valid until this date inclusive
          format: date
          nullable: true
        external_reference_uri:
          type: string
          description: A link (URI) to an external definition/documentation
          format: url
          nullable: true
        comment:
          nullable: true
          allOf:
          - $ref: "#/components/schemas/LanguageStringType"
        related_variable_definition_uris:
          type: array
          description: "Link(s) to related definitions of variables - a list of one\
            \ or more definitions. For example for a variable after-tax income it\
            \ could be relevant to link to definitions of income from work, property\
            \ income etc."
          nullable: true
          items:
            type: string
            format: url
        owner:
          $ref: "#/components/schemas/Owner"
        contact:
          $ref: "#/components/schemas/Contact"
        created_at:
          type: string
          description: The timestamp at which this variable definition was first created.
          format: date-time
        created_by:
          type: string
          description: The user who created this variable definition.
        last_updated_at:
          type: string
          description: The timestamp at which this variable definition was last modified.
          format: date-time
        last_updated_by:
          type: string
          description: The user who last modified this variable definition.
      description: Complete view For internal users who need all details while maintaining
        variable definitions.
    Contact:
      required:
      - email
      - title
      type: object
      properties:
        title:
          $ref: "#/components/schemas/LanguageStringType"
        email:
          type: string
          format: email
      description: Contact details
...
    LanguageStringType:
      type: object
      properties:
        nb:
          type: string
          description: Norwegian Bokmål
          nullable: true
        nn:
          type: string
          description: Norwegian Nynorsk
          nullable: true
        en:
          type: string
          description: English
          nullable: true
      description: "Language string type Represents one text, with translations for\
        \ the languages in \\[SupportedLanguages\\]. All fields are nullable to allow\
        \ for flexibility for maintainers."
    Owner:
      required:
      - groups
      - team
      type: object
      properties:
        team:
          minLength: 1
          type: string
          description: The Dapla team with responsibility for this variable definition.
        groups:
          minItems: 1
          type: array
          description: The groups with permission to modify this variable definition.
          items:
            minLength: 1
            type: string
      description: "Owner of the definition, i.e. responsible Dapla team (statistics\
        \ team) and information about access management groups."
      example:
        team: play-obr-b
        groups:
        - play-obr-b-developers
...
    VardokIdResponse:
      required:
      - vardok_id
      type: object
      properties:
        vardok_id:
          type: string
    VardokVardefIdPairResponse:
      required:
      - vardef_id
      - vardok_id
      type: object
      properties:
        vardok_id:
          type: string
        vardef_id:
          type: string
    VariableStatus:
      type: string
      description: Status of the life cycle of the variable
      enum:
      - DRAFT
      - PUBLISHED_INTERNAL
      - PUBLISHED_EXTERNAL
  securitySchemes:
    labid_token:
      type: http
      description: A token granted by Statistics Norway's LabID instance. May be obtained
        from a <a href=https://lab.dapla.ssb.no>Dapla Lab</a> service.
      scheme: bearer
      bearerFormat: JWT